+++ packages/instant-meilisearch/__tests__/overridden-meilisearch-parameters.test.ts
@@ -4,12 +4,8 @@ import { dataset, meilisearchClient, type Movies } from './assets/utils.js'
 
 describe('InstantMeiliSearch overridden parameters', () => {
   beforeAll(async () => {
-    const deleteTask = await meilisearchClient.deleteIndex('movies')
-    await meilisearchClient.waitForTask(deleteTask.taskUid)
-    const documentsTask = await meilisearchClient
-      .index('movies')
-      .addDocuments(dataset)
-    await meilisearchClient.index('movies').waitForTask(documentsTask.taskUid)
+    await meilisearchClient.deleteIndex('movies').waitTask()
+    await meilisearchClient.index('movies').addDocuments(dataset).waitTask()
   })
 
   test('instantiating with, and changing overridden Meilisearch parameters', async () => {